requirements.txt
dj-database-url==0.4.2
Django==1.11.4
django-extensions==1.9.0
django-model-utils==3.0.0
django-widget-tweaks==1.4.1
dj-static==0.0.6
gunicorn==19.7.1
Pillow==4.2.1
psycopg2==2.7.3.1
python-decouple==3.1
ipython==6.1.0
model-mommy==1.3.2
pytz
toolz==0.8.2
django-measurement==2.4.0
djangorestframework==3.7.3
